Fix an error with multiple cluster shards
ClusterInfo Hash is a required parameter. For a given cluster matching an
ClusterHealthCheck this is set by the Sveltos deployment matching the
cluster shard.

Other deployment will simply skip such ClusterInfo. Since this is a
required field, if deployment matching the shard it has not processed the
ClusterHealthCheck instance, any other Sveltos deployment will set it to
empty if not set already.